Source: https://github.com/msupply-foundation/open-msupply/commit/66e111b23943c6bf5c31e8bcd18e733319980acb

+++ client/packages/system/src/IndicatorsDemographics/DetailView/PercentageColumn.tsx
@@ -43,7 +43,7 @@ export const percentageColumn = <
             const updatedRowData = { ...rowData };
             column.setter({
               ...updatedRowData,
-              percentage: parseInt(e.target.value),
+              percentage: parseFloat(e.target.value),
             });
           }}
         />
